Football Cup
It is the final day of La Liga. Chef has a certain criteria for assessing football matches.
In particular, he only likes a match if:
The match ends in a draw, and at least one goal has been scored by either team.
Given the goals scored by both the teams as 
X and Y respectively, determine whether Chef will like the match or not.